........
  r3501 | stefan | 2007-01-08 16:56:54 -0500 (Mon, 08 Jan 2007) | 2 lines
  
  Whitespace cleanup.
........
  r3520 | timl | 2007-01-25 23:16:37 -0500 (Thu, 25 Jan 2007) | 1 line
  
  remove some remaining references to scipy. fixes #428
........
  r3521 | rkern | 2007-01-26 19:35:17 -0500 (Fri, 26 Jan 2007) | 1 line
  
  Generalize the setting of MACOSX_DEPLOYMENT_TARGET.
........
  r3523 | rkern | 2007-01-30 14:18:06 -0500 (Tue, 30 Jan 2007) | 1 line
  
  Revert semi-intelligent handling of MACOSX_DEPLOYMENT_TARGET to work around broken MacPorts Python.
........
  r3538 | edschofield | 2007-02-04 19:24:42 -0500 (Sun, 04 Feb 2007) | 2 lines
  
  Fixed detection of Core2 CPUs (scipy ticket #349)
........
  r3549 | stefan | 2007-02-18 16:02:46 -0500 (Sun, 18 Feb 2007) | 2 lines
  
  Fix docstrings for loading with DocFileSuite.
........
  r3563 | pearu | 2007-02-28 08:12:06 -0500 (Wed, 28 Feb 2007) | 1 line
  
  Added support for xml Fortran V10 compiler (needs testing).
........
  r3564 | pearu | 2007-02-28 13:32:18 -0500 (Wed, 28 Feb 2007) | 1 line
  
  Fixing xlf compiler for AIX with 9.x and 10.x version support (needs testing and checking that xlf 8.x compilers are still detected)
........
  r3568 | pearu | 2007-03-02 11:49:00 -0500 (Fri, 02 Mar 2007) | 1 line
  
  Applying info from Hans-Joachim Ehlers to improve xlf compiler support (on AIX using lslpp to determine compiler version).
........
  r3569 | timl | 2007-03-04 05:47:08 -0500 (Sun, 04 Mar 2007) | 1 line
  
  add a delete flag to make_svn_version_py to allow the __svn_version.py to not be deleted
........
  r3598 | rkern | 2007-03-25 19:12:00 -0400 (Sun, 25 Mar 2007) | 1 line
  
  Make the gfortran version matching regex accept gfortran 4.3 on OS X
........
  r3615 | cookedm | 2007-03-29 17:37:58 -0400 (Thu, 29 Mar 2007) | 1 line
  
  cpuinfo: Nocona is a PentiumIV, not 686. Fixes #438
........
  r3636 | cookedm | 2007-04-01 16:32:07 -0400 (Sun, 01 Apr 2007) | 2 lines
  
  Make numpy.get_include() and numpy.get_numarray_include() not pull in numpy.distutils.
........
  r3658 | cookedm | 2007-04-04 08:56:47 -0400 (Wed, 04 Apr 2007) | 2 lines
  
  Fix for making Fortran extensions on AIX.
........
  r3660 | rkern | 2007-04-04 13:25:29 -0400 (Wed, 04 Apr 2007) | 1 line
  
  Fix #491
........
  r3664 | cookedm | 2007-04-04 14:10:30 -0400 (Wed, 04 Apr 2007) | 2 lines
  
  Really fix linking with Fortran on AIX
........
  r3715 | rkern | 2007-04-15 00:38:59 -0400 (Sun, 15 Apr 2007) | 1 line
  
  Fix command dependencies for build_ext. This should fix a number of problems where the full 'build_src build_clib build_ext' chain of commands needed to be specified.
........
  r3716 | rkern | 2007-04-15 00:46:07 -0400 (Sun, 15 Apr 2007) | 1 line
  
  Add another place to check for configuration, namely ~/.numpy-site.cfg, to help automatic installation where one can't edit the source checkout. I've added the alternate name to make it hidden because that's just considerate.
........
  r3723 | cookedm | 2007-04-22 16:57:58 -0400 (Sun, 22 Apr 2007) | 9 lines
  
  Better version handling for gnu and intel Fortran compilers
  
  - gnu compilers check if the version is >= 4, in which case it's gfortran
  - add a test file for gnu compiler check
  - simplify version matching on intel compilers to be more flexible
  - add FCompiler.find_executables so that subclasses can find executables
    at .customize() time, instead of when the class is created.
........
  r3724 | cookedm | 2007-04-22 17:12:57 -0400 (Sun, 22 Apr 2007) | 14 lines
  
  Some distutils work:
  
  - Add better support for C++ in numpy.distutils. Instead of munging the
    C compiler command, build_clib and build_ext call the new
    Compiler.cxx_compiler() method to get a version of the compiler suitable for
    C++ (this also takes care of the special needs of AIX).
  - If config_fc is specified in the Extension definition, merge that info
    instead of replacing it (otherwise, the name of the Fortran compiler is
    overwritten). This is done at the key level (ex., compiler options are
    replaced instead of appended).
  - clean up compiler.py a bit
  - clean up linking in build_ext
........
  r3726 | oliphant | 2007-04-24 15:15:33 -0400 (Tue, 24 Apr 2007) | 1 line
  
  Add patch to system_info for building with MKL on Win32 #504
........
  r3731 | rkern | 2007-04-30 12:29:51 -0400 (Mon, 30 Apr 2007) | 1 line
  
  Fix typo.
........

+++ misc_util.py
@@ -962,9 +962,9 @@ class Configuration(object):
         return
 
     def add_numarray_include_dirs(self):
-	import numpy.numarray.util as nnu
-	self.add_include_dirs(*nnu.get_numarray_include_dirs())
-	
+        import numpy.numarray.util as nnu
+        self.add_include_dirs(*nnu.get_numarray_include_dirs())
+
     def add_headers(self,*files):
         """ Add installable headers to configuration.
         Argument(s) can be either
@@ -1294,7 +1294,7 @@ class Configuration(object):
 
         return version
 
-    def make_svn_version_py(self):
+    def make_svn_version_py(self, delete=True):
         """ Generate package __svn_version__.py file from SVN revision number,
         it will be removed after python exits but will be available
         when sdist, etc commands are executed.
@@ -1302,29 +1302,31 @@ class Configuration(object):
         If __svn_version__.py existed before, nothing is done.
         """
         target = njoin(self.local_path,'__svn_version__.py')
-        if os.path.isfile(target):
+        revision = self._get_svn_revision(self.local_path)
+        if os.path.isfile(target) or revision is None:
             return
-        def generate_svn_version_py():
-            if not os.path.isfile(target):
-                revision = self._get_svn_revision(self.local_path)
-                assert revision is not None,'hmm, why I am not inside SVN tree???'
-                version = str(revision)
-                self.info('Creating %s (version=%r)' % (target,version))
-                f = open(target,'w')
-                f.write('version = %r\n' % (version))
-                f.close()
+        else:
+            def generate_svn_version_py():
+                if not os.path.isfile(target):
+                    version = str(revision)
+                    self.info('Creating %s (version=%r)' % (target,version))
+                    f = open(target,'w')
+                    f.write('version = %r\n' % (version))
+                    f.close()
 
-            import atexit
-            def rm_file(f=target,p=self.info):
-                try: os.remove(f); p('removed '+f)
-                except OSError: pass
-                try: os.remove(f+'c'); p('removed '+f+'c')
-                except OSError: pass
-            atexit.register(rm_file)
+                import atexit
+                def rm_file(f=target,p=self.info):
+                    if delete:
+                        try: os.remove(f); p('removed '+f)
+                        except OSError: pass
+                        try: os.remove(f+'c'); p('removed '+f+'c')
+                        except OSError: pass
 
-            return target
+                atexit.register(rm_file)
 
-        self.add_data_files(('', generate_svn_version_py()))
+                return target
+
+            self.add_data_files(('', generate_svn_version_py()))
 
     def make_config_py(self,name='__config__'):
         """ Generate package __config__.py file containing system_info
@@ -1360,14 +1362,7 @@ def get_numpy_include_dirs():
     include_dirs = Configuration.numpy_include_dirs[:]
     if not include_dirs:
         import numpy
-        if numpy.show_config is None:
-            # running from numpy_core source directory
-            include_dirs.append(njoin(os.path.dirname(numpy.__file__),
-                                             'core', 'include'))
-        else:
-            # using installed numpy core headers
-            import numpy.core as core
-            include_dirs.append(njoin(os.path.dirname(core.__file__), 'include'))
+        include_dirs = [ numpy.get_include() ]
     # else running numpy/core/setup.py
     return include_dirs
